Tutorial 2: Creating the tile class
In this tutorial we create our first class, a Tile class.  In this class we first define a variable to give each tile generated a unique ID, we then create an array to track of all the adjacent tiles.  We finish this tutorial by creating an enum to track the indexes in the array and which side they correspond to.