Tutorial 5: Connect the map tester to the map class
In the previous video we created our RandomMapTester class and we also made a method that allowed us to create a new map. The only problem is that we don't actually have a way of calling it. If you click on the RandomMapTester instance, in the scene, you'll see that there's no way to actually call that method. We're gonna fix that by creating a custom editor and in that editor, we'll create a button that'll allow us to call the method to create a new map. Let's go into our editor folder and create a new C# script.